Move world-age DI test and drop second-order ext suite
