--- test/ext/differentiation_interface_second_order/Project.toml
@@ -1,5 +1,0 @@
-[deps]
-DifferentiationInterface = "a0c0ee7d-e4b9-4e03-894e-1c5f64a51d63"
-DifferentiationInterfaceTest = "a82114a7-5aa3-49a8-9643-716bb13727a3"
-Mooncake = "da2b9cff-9c12-43a0-ae48-6db2b0edb7d6"
-Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
